CMA Separation Step

First CMA step: decomposes the the pure "individual" features of marginal distributions and the pure "joint" information available in the copulas. distributions from the pure "joint" information available in the copulas.

cma_separation(x, p = NULL)

# S3 method for default
cma_separation(x, p = NULL)

# S3 method for matrix
cma_separation(x, p = NULL)

# S3 method for xts
cma_separation(x, p = NULL)

# S3 method for data.frame
cma_separation(x, p = NULL)

# S3 method for tbl
cma_separation(x, p = NULL)

Arguments

x
A rectangular (non-tidy) data structure.

p
A probability vector.

Value

An S3 list of the cma class that contains two elements: marginal and copula.

Details

Arguments x and p must have the same size.

See also

Examples

set.seed(123)
margins <- matrix(stats::rnorm(30), ncol = 3)
cma <- cma_separation(x = margins)
cma

#> # CMA Decomposition
#> marginal: << dim 10 x 3 >>
#> cdf     : << dim 10 x 3 >>
#> copula  : << dim 10 x 3 >>

# Access the elements with `$`
cma$copula

#>              ..1        ..2        ..3
#>  [1,] 0.27272727 0.81818182 0.27272727
#>  [2,] 0.45454545 0.45454545 0.63636364
#>  [3,] 0.81818182 0.54545455 0.36363636
#>  [4,] 0.54545455 0.36363636 0.45454545
#>  [5,] 0.63636364 0.18181818 0.54545455
#>  [6,] 0.90909091 0.90909091 0.09090909
#>  [7,] 0.72727273 0.63636364 0.81818182
#>  [8,] 0.09090909 0.09090909 0.72727273
#>  [9,] 0.18181818 0.72727273 0.18181818
#> [10,] 0.36363636 0.27272727 0.90909091

cma$marginal

#>               ..1        ..2        ..3
#>  [1,] -0.56047565  1.2240818 -1.0678237
#>  [2,] -0.23017749  0.3598138 -0.2179749
#>  [3,]  1.55870831  0.4007715 -1.0260044
#>  [4,]  0.07050839  0.1106827 -0.7288912
#>  [5,]  0.12928774 -0.5558411 -0.6250393
#>  [6,]  1.71506499  1.7869131 -1.6866933
#>  [7,]  0.46091621  0.4978505  0.8377870
#>  [8,] -1.26506123 -1.9666172  0.1533731
#>  [9,] -0.68685285  0.7013559 -1.1381369
#> [10,] -0.44566197 -0.4727914  1.2538149

cma$cdf

#>              ..1        ..2        ..3
#>  [1,] 0.09090909 0.09090909 0.09090909
#>  [2,] 0.18181818 0.18181818 0.18181818
#>  [3,] 0.27272727 0.27272727 0.27272727
#>  [4,] 0.36363636 0.36363636 0.36363636
#>  [5,] 0.45454545 0.45454545 0.45454545
#>  [6,] 0.54545455 0.54545455 0.54545455
#>  [7,] 0.63636364 0.63636364 0.63636364
#>  [8,] 0.72727273 0.72727273 0.72727273
#>  [9,] 0.81818182 0.81818182 0.81818182
#> [10,] 0.90909091 0.90909091 0.90909091