Moment Matching for the Normal Distribution

Generates scenarios from the normal distribution that exactly matches the first two moments.

match_normal(mu, sigma, n)

Arguments

mu
A double with the location parameter.

sigma
A variance-covariance matrix the dispersion parameter.

n
An integer with the number of scenarios to be generated.

Value

#TODO

See also

Examples

x     <- diff(log(EuStockMarkets))
mu    <- colMeans(x)
sigma <- cov(x)
match_normal(mu, sigma, 10)

#>               [,1]         [,2]         [,3]         [,4]
#>  [1,] -0.002028885 -0.009268531 -0.009977048 -0.001287507
#>  [2,] -0.002958265  0.005832372 -0.003125622  0.010325578
#>  [3,] -0.005787895 -0.010920803 -0.010984207 -0.003051279
#>  [4,]  0.022208973  0.012842283  0.013172433  0.012235431
#>  [5,] -0.001382765  0.005125090 -0.013513566 -0.007588792
#>  [6,]  0.003332968  0.010904330  0.010851156  0.002151477
#>  [7,]  0.004262349 -0.004196573  0.003999730 -0.009461608
#>  [8,]  0.007091978  0.012556602  0.011858314  0.003915249
#>  [9,] -0.020904889 -0.011206484 -0.012298325 -0.011371461
#> [10,]  0.002686849 -0.003489291  0.014387674  0.008452762

# correlation structure is similar
cor(x)

#>            DAX       SMI       CAC      FTSE
#> DAX  1.0000000 0.7031219 0.7344304 0.6394674
#> SMI  0.7031219 1.0000000 0.6160454 0.5847791
#> CAC  0.7344304 0.6160454 1.0000000 0.6485679
#> FTSE 0.6394674 0.5847791 0.6485679 1.0000000

cor(match_normal(mu, sigma, 10000))

#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 1.0000000 0.7031219 0.7344304 0.6394674
#> [2,] 0.7031219 1.0000000 0.6160454 0.5847791
#> [3,] 0.7344304 0.6160454 1.0000000 0.6485679
#> [4,] 0.6394674 0.5847791 0.6485679 1.0000000

# as the location parameters
mu

#>          DAX          SMI          CAC         FTSE 
#> 0.0006520417 0.0008178997 0.0004370540 0.0004319851 

colMeans(match_normal(mu, sigma, 10000))

#> [1] 0.0006520417 0.0008178997 0.0004370540 0.0004319851