Moment Matching for the t Distribution

Generates scenarios from the Student-t distribution that exactly matches the first two moments.

match_t(mu, sigma, nu, n, stochastic = FALSE)

Arguments

mu
A double with the location parameter.

sigma
A variance-covariance matrix the dispersion parameter.

nu
An integer with the degrees of freedom.

n
An integer with the number of scenarios to be generated.

stochastic
A logical flag indicating whether (or not) the stochastic representation should be used. The default if FALSE.

Value

# TODO

See also

Examples

x     <- diff(log(EuStockMarkets))
mu    <- colMeans(x)
sigma <- cov(x)
nu    <- 5
match_t(mu, sigma, nu, 10)

#>                [,1]          [,2]          [,3]         [,4]
#>  [1,]  0.0164244658  0.0129117413  0.0018572518 -0.005524058
#>  [2,]  0.0061213793  0.0098333912  0.0126673165  0.006940157
#>  [3,]  0.0079027877 -0.0008374713  0.0013785530  0.006892041
#>  [4,] -0.0158987238 -0.0141422369 -0.0172534518 -0.012833765
#>  [5,]  0.0040513031 -0.0033150652  0.0090505799 -0.001425389
#>  [6,] -0.0078050085 -0.0063880245  0.0001590019  0.003690972
#>  [7,] -0.0116451611 -0.0207659363 -0.0265579023 -0.014174665
#>  [8,] -0.0076523066  0.0024308232 -0.0005875199 -0.005939102
#>  [9,]  0.0209371011  0.0190196680  0.0226198657  0.017394360
#> [10,] -0.0002958565  0.0016613539 -0.0019338522  0.001042208

cor(x)

#>            DAX       SMI       CAC      FTSE
#> DAX  1.0000000 0.7031219 0.7344304 0.6394674
#> SMI  0.7031219 1.0000000 0.6160454 0.5847791
#> CAC  0.7344304 0.6160454 1.0000000 0.6485679
#> FTSE 0.6394674 0.5847791 0.6485679 1.0000000

cor(match_t(mu, sigma, 5, 10000))

#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 1.0000000 0.6866174 0.7307546 0.6310661
#> [2,] 0.6866174 1.0000000 0.6140364 0.5748925
#> [3,] 0.7307546 0.6140364 1.0000000 0.6404942
#> [4,] 0.6310661 0.5748925 0.6404942 1.0000000

mu

#>          DAX          SMI          CAC         FTSE 
#> 0.0006520417 0.0008178997 0.0004370540 0.0004319851 

colMeans(match_t(mu, sigma, nu, 10000))

#> [1] 0.0004850539 0.0005212520 0.0004975195 0.0005516226