Panic-Copula for Stress-Testing

This function generates a large dimensional panic copula that can be further used for stress-testing and panic-aware portfolio optimization.

panic_copula(
  x,
  n = 10000,
  panic_cor = 0.99,
  panic_prob = 0.02,
  dist = c("normal", "t")
)

Arguments

x
A rectangular (non-tidy) data structure.

n
An integer with the number of scenarios to be generated.

panic_cor
A numeric value for the correlation in panic markets.

panic_prob
A numeric value with the probability in which panic markets can be triggered.

dist
A character with normal or t. The default is normal

Value

An object of the panic class.

Examples

x <- diff(log(EuStockMarkets))
panic_copula(x = x, n = 20, panic_cor = 0.99, panic_prob = 0.02)

#> $simulation
#>               [,1]         [,2]         [,3]        [,4]
#>  [1,] -0.088496367 -0.131833928 -0.128440788 -0.20003832
#>  [2,]  0.001623434 -0.025778094  0.063361703  0.14923580
#>  [3,]  0.150754525  0.096867104  0.153977167  0.12094491
#>  [4,]  0.045836655  0.002349218 -0.039259486  0.02956334
#>  [5,] -0.040632846  0.077567823  0.017330575 -0.07241235
#>  [6,]  0.121901636  0.194582299  0.047155289  0.07858886
#>  [7,] -0.026514189 -0.040867341 -0.010890400 -0.05662751
#>  [8,] -0.071115642  0.061119132  0.079972768  0.00129530
#>  [9,] -0.158110086 -0.109334815 -0.106541450 -0.10822715
#> [10,]  0.098044504 -0.055622750  0.194582299  0.01471196
#> [11,]  0.080127723  0.149513582  0.124298151  0.19458230
#> [12,] -0.012879153  0.015832212 -0.069772486 -0.13048437
#> [13,] -0.200879068 -0.200879068 -0.200879068 -0.15954701
#> [14,] -0.055642359 -0.012250234  0.030969324 -0.04202945
#> [15,]  0.030448945 -0.089057743 -0.024593826  0.06031718
#> [16,] -0.130762893 -0.159289458 -0.055089021 -0.08997472
#> [17,]  0.016039461  0.030493159  0.003491557  0.04468514
#> [18,]  0.061576893 -0.071137542 -0.086600109 -0.01247461
#> [19,]  0.194582299  0.119557229  0.099836208  0.09826040
#> [20,] -0.106388462  0.045840791 -0.154784061 -0.02683407
#> 
#> $p
#>             [,1]
#>  [1,] 0.04531510
#>  [2,] 0.05199406
#>  [3,] 0.05238225
#>  [4,] 0.05235466
#>  [5,] 0.04947864
#>  [6,] 0.05492518
#>  [7,] 0.04862435
#>  [8,] 0.04937134
#>  [9,] 0.04735651
#> [10,] 0.04804812
#> [11,] 0.05536423
#> [12,] 0.04825235
#> [13,] 0.04462584
#> [14,] 0.04792000
#> [15,] 0.05070543
#> [16,] 0.04567733
#> [17,] 0.05159628
#> [18,] 0.05081562
#> [19,] 0.05297762
#> [20,] 0.05221507
#>