Moment Matching for the Normal Distribution

Generates scenarios from the normal distribution that exactly matches the first two moments.

match_normal(mu, sigma, n)

Arguments

mu
A double with the location parameter.

sigma
A variance-covariance matrix the dispersion parameter.

n
An integer with the number of scenarios to be generated.

Value

#TODO

See also

Examples

x     <- diff(log(EuStockMarkets))
mu    <- colMeans(x)
sigma <- cov(x)
match_normal(mu, sigma, 10)

#>                [,1]          [,2]          [,3]         [,4]
#>  [1,] -0.0089636854  0.0032513142 -0.0157357512 -0.003323587
#>  [2,]  0.0163946294  0.0184675242  0.0181355838  0.014725172
#>  [3,]  0.0020231263  0.0045767184  0.0012453326  0.007933306
#>  [4,] -0.0087117626 -0.0088728310 -0.0027547439  0.005604906
#>  [5,]  0.0106861930 -0.0007137982 -0.0043339569  0.004330540
#>  [6,]  0.0102677689 -0.0016155149  0.0166098592  0.004187557
#>  [7,] -0.0150905459 -0.0168317249 -0.0172614758 -0.013861201
#>  [8,] -0.0007190428 -0.0029409191 -0.0003712246 -0.007069336
#>  [9,]  0.0100158461  0.0105086303  0.0036288519 -0.004740936
#> [10,] -0.0093821095  0.0023495975  0.0052080649 -0.003466570

# correlation structure is similar
cor(x)

#>            DAX       SMI       CAC      FTSE
#> DAX  1.0000000 0.7031219 0.7344304 0.6394674
#> SMI  0.7031219 1.0000000 0.6160454 0.5847791
#> CAC  0.7344304 0.6160454 1.0000000 0.6485679
#> FTSE 0.6394674 0.5847791 0.6485679 1.0000000

cor(match_normal(mu, sigma, 10000))

#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 1.0000000 0.7031219 0.7344304 0.6394674
#> [2,] 0.7031219 1.0000000 0.6160454 0.5847791
#> [3,] 0.7344304 0.6160454 1.0000000 0.6485679
#> [4,] 0.6394674 0.5847791 0.6485679 1.0000000

# as the location parameters
mu

#>          DAX          SMI          CAC         FTSE 
#> 0.0006520417 0.0008178997 0.0004370540 0.0004319851 

colMeans(match_normal(mu, sigma, 10000))

#> [1] 0.0006520417 0.0008178997 0.0004370540 0.0004319851