Moment Matching for the Normal Distribution

Generates scenarios from the normal distribution that exactly matches the first two moments.

match_normal(mu, sigma, n)

Arguments

mu
A double with the location parameter.

sigma
A variance-covariance matrix the dispersion parameter.

n
An integer with the number of scenarios to be generated.

Value

#TODO

See also

Examples

x     <- diff(log(EuStockMarkets))
mu    <- colMeans(x)
sigma <- cov(x)
match_normal(mu, sigma, 10)

#> # A tibble: 10 × 4
#>         DAX        SMI      CAC     FTSE
#>       <dbl>      <dbl>    <dbl>    <dbl>
#>  1 -0.00912 -0.0166    -0.0172  -0.0105 
#>  2 -0.00162 -0.00635    0.0104   0.00411
#>  3 -0.0198  -0.00666   -0.0113  -0.00879
#>  4  0.00256 -0.00302    0.00471  0.00283
#>  5 -0.00235 -0.0000732 -0.00606  0.0101 
#>  6  0.0104   0.0183     0.0181   0.0113 
#>  7  0.00292  0.00799   -0.00949 -0.00324
#>  8  0.0211   0.00829    0.0122   0.00965
#>  9 -0.00126  0.00466   -0.00383 -0.00196
#> 10  0.00365  0.00171    0.00693 -0.00924

# correlation structure is similar
cor(x)

#>            DAX       SMI       CAC      FTSE
#> DAX  1.0000000 0.7031219 0.7344304 0.6394674
#> SMI  0.7031219 1.0000000 0.6160454 0.5847791
#> CAC  0.7344304 0.6160454 1.0000000 0.6485679
#> FTSE 0.6394674 0.5847791 0.6485679 1.0000000

cor(match_normal(mu, sigma, 10000))

#>            DAX       SMI       CAC      FTSE
#> DAX  1.0000000 0.7031219 0.7344304 0.6394674
#> SMI  0.7031219 1.0000000 0.6160454 0.5847791
#> CAC  0.7344304 0.6160454 1.0000000 0.6485679
#> FTSE 0.6394674 0.5847791 0.6485679 1.0000000

# as the location parameters
mu

#>          DAX          SMI          CAC         FTSE 
#> 0.0006520417 0.0008178997 0.0004370540 0.0004319851 

colMeans(match_normal(mu, sigma, 10000))

#>          DAX          SMI          CAC         FTSE 
#> 0.0006520417 0.0008178997 0.0004370540 0.0004319851