Moment Matching for the Normal Distribution

Generates scenarios from the normal distribution that exactly matches the first two moments.

match_normal(mu, sigma, n)

Arguments

mu
A double with the location parameter.

sigma
A variance-covariance matrix the dispersion parameter.

n
An integer with the number of scenarios to be generated.

Value

#TODO

See also

Examples

x     <- diff(log(EuStockMarkets))
mu    <- colMeans(x)
sigma <- cov(x)
match_normal(mu, sigma, 10)

#>                [,1]         [,2]         [,3]         [,4]
#>  [1,] -0.0112865717 -0.005044912 -0.003490341  0.005432531
#>  [2,] -0.0023989421 -0.009073385  0.008770385 -0.003306144
#>  [3,]  0.0007940403  0.005942585  0.004525653  0.005305001
#>  [4,] -0.0176339341 -0.014304201 -0.021969634 -0.014251958
#>  [5,]  0.0073078741 -0.005559045 -0.001743746  0.006619362
#>  [6,]  0.0125906552  0.006680711  0.004364449 -0.004568560
#>  [7,]  0.0037030256  0.010709185 -0.007896277  0.004170114
#>  [8,]  0.0005100432 -0.004306786 -0.003651545 -0.004441031
#>  [9,]  0.0189380176  0.015940001  0.022843742  0.015115928
#> [10,] -0.0060037906  0.007194844  0.002617854 -0.005755392

# correlation structure is similar
cor(x)

#>            DAX       SMI       CAC      FTSE
#> DAX  1.0000000 0.7031219 0.7344304 0.6394674
#> SMI  0.7031219 1.0000000 0.6160454 0.5847791
#> CAC  0.7344304 0.6160454 1.0000000 0.6485679
#> FTSE 0.6394674 0.5847791 0.6485679 1.0000000

cor(match_normal(mu, sigma, 10000))

#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 1.0000000 0.7031219 0.7344304 0.6394674
#> [2,] 0.7031219 1.0000000 0.6160454 0.5847791
#> [3,] 0.7344304 0.6160454 1.0000000 0.6485679
#> [4,] 0.6394674 0.5847791 0.6485679 1.0000000

# as the location parameters
mu

#>          DAX          SMI          CAC         FTSE 
#> 0.0006520417 0.0008178997 0.0004370540 0.0004319851 

colMeans(match_normal(mu, sigma, 10000))

#> [1] 0.0006520417 0.0008178997 0.0004370540 0.0004319851