Moment Matching for the t Distribution

Generates scenarios from the Student-t distribution that exactly matches the first two moments.

match_t(mu, sigma, nu, n, stochastic = FALSE)

Arguments

mu
A double with the location parameter.

sigma
A variance-covariance matrix the dispersion parameter.

nu
An integer with the degrees of freedom.

n
An integer with the number of scenarios to be generated.

stochastic
A logical flag indicating whether (or not) the stochastic representation should be used. The default if FALSE.

Value

# TODO

See also

Examples

x     <- diff(log(EuStockMarkets))
mu    <- colMeans(x)
sigma <- cov(x)
nu    <- 5
match_t(mu, sigma, nu, 10)

#>               [,1]          [,2]          [,3]          [,4]
#>  [1,] -0.013723199 -0.0160947827 -0.0173432173 -0.0107354720
#>  [2,]  0.010924016  0.0022454783  0.0142670774  0.0005955867
#>  [3,]  0.009199803  0.0041302108 -0.0013335713  0.0035418288
#>  [4,] -0.001340425  0.0203706534  0.0001311610  0.0078268218
#>  [5,] -0.003017733  0.0055484771 -0.0089885369 -0.0184503351
#>  [6,]  0.020799996  0.0234118259  0.0258319496  0.0159621935
#>  [7,] -0.004457350 -0.0002262351 -0.0060763984  0.0005728093
#>  [8,] -0.015866937 -0.0056515771  0.0037253900 -0.0045571727
#>  [9,]  0.002394152 -0.0187814417  0.0009477231 -0.0064520493
#> [10,]  0.003759300 -0.0036649257  0.0085450217  0.0155706159

cor(x)

#>            DAX       SMI       CAC      FTSE
#> DAX  1.0000000 0.7031219 0.7344304 0.6394674
#> SMI  0.7031219 1.0000000 0.6160454 0.5847791
#> CAC  0.7344304 0.6160454 1.0000000 0.6485679
#> FTSE 0.6394674 0.5847791 0.6485679 1.0000000

cor(match_t(mu, sigma, 5, 10000))

#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 1.0000000 0.7064170 0.7333618 0.6407409
#> [2,] 0.7064170 1.0000000 0.6267651 0.5998516
#> [3,] 0.7333618 0.6267651 1.0000000 0.6497041
#> [4,] 0.6407409 0.5998516 0.6497041 1.0000000

mu

#>          DAX          SMI          CAC         FTSE 
#> 0.0006520417 0.0008178997 0.0004370540 0.0004319851 

colMeans(match_t(mu, sigma, nu, 10000))

#> [1] 0.0006138829 0.0005287249 0.0006790491 0.0006294690