Moment Matching for the t Distribution

Generates scenarios from the Student-t distribution that exactly matches the first two moments.

match_t(mu, sigma, nu, n, stochastic = FALSE)

Arguments

mu
A double with the location parameter.

sigma
A variance-covariance matrix the dispersion parameter.

nu
An integer with the degrees of freedom.

n
An integer with the number of scenarios to be generated.

stochastic
A logical flag indicating whether (or not) the stochastic representation should be used. The default if TRUE.

Value

A tibble in the wide format.

See also

Examples

x     <- diff(log(EuStockMarkets))
mu    <- colMeans(x)
sigma <- cov(x)
nu    <- 5
match_t(mu, sigma, nu, 10)

#> # A tibble: 10 × 4
#>         DAX      SMI       CAC       FTSE
#>       <dbl>    <dbl>     <dbl>      <dbl>
#>  1 -0.00111 -0.0151  -0.0182   -0.00821  
#>  2  0.00664  0.0107  -0.000563  0.00126  
#>  3 -0.0139  -0.00584 -0.0101   -0.00754  
#>  4  0.0132   0.00566  0.0186    0.00208  
#>  5 -0.00244 -0.00236 -0.00266  -0.0144   
#>  6  0.00194  0.0104   0.0129    0.00597  
#>  7 -0.00741 -0.0132   0.00230  -0.000468 
#>  8  0.0279   0.0132   0.0205    0.0165   
#>  9 -0.00342 -0.00111 -0.00515   0.0000343
#> 10  0.00351  0.00287  0.00370   0.0134   

cor(x)

#>            DAX       SMI       CAC      FTSE
#> DAX  1.0000000 0.7031219 0.7344304 0.6394674
#> SMI  0.7031219 1.0000000 0.6160454 0.5847791
#> CAC  0.7344304 0.6160454 1.0000000 0.6485679
#> FTSE 0.6394674 0.5847791 0.6485679 1.0000000

cor(match_t(mu, sigma, 5, 100000))

#>            DAX       SMI       CAC      FTSE
#> DAX  1.0000000 0.7009054 0.7328550 0.6381210
#> SMI  0.7009054 1.0000000 0.6124058 0.5800144
#> CAC  0.7328550 0.6124058 1.0000000 0.6479002
#> FTSE 0.6381210 0.5800144 0.6479002 1.0000000

mu

#>          DAX          SMI          CAC         FTSE 
#> 0.0006520417 0.0008178997 0.0004370540 0.0004319851 

colMeans(match_t(mu, sigma, nu, 100000))

#>          DAX          SMI          CAC         FTSE 
#> 0.0005937165 0.0006112781 0.0005875299 0.0005845531