Panic-Copula for Stress-Testing

This function generates a large dimensional panic copula that can be further used for stress-testing and panic-aware portfolio optimization.

panic_copula(
  x,
  n = 10000,
  panic_cor = 0.99,
  panic_prob = 0.02,
  dist = c("normal", "t")
)

Arguments

x
A rectangular (non-tidy) data structure.

n
An integer with the number of scenarios to be generated.

panic_cor
A numeric value for the correlation in panic markets.

panic_prob
A numeric value with the probability in which panic markets can be triggered.

dist
A character with normal or t. The default is normal

Value

An object of the panic class.

Examples

x <- diff(log(EuStockMarkets))
panic_copula(x = x, n = 20, panic_cor = 0.99, panic_prob = 0.02)

#> $simulation
#>                [,1]          [,2]          [,3]          [,4]
#>  [1,] -0.0078161766 -0.0104560352 -0.0121481277 -1.364893e-02
#>  [2,]  0.0001433849 -0.0020445166  0.0059928476  1.018260e-02
#>  [3,]  0.0133149419  0.0076827405  0.0145633977  8.252263e-03
#>  [4,]  0.0040483853  0.0001863216 -0.0037132227  2.017154e-03
#>  [5,] -0.0035887744  0.0061520726  0.0016391525 -4.940810e-03
#>  [6,]  0.0107665970  0.0154327448  0.0044600199  5.362243e-03
#>  [7,] -0.0023417863 -0.0032412776 -0.0010300308 -3.863785e-03
#>  [8,] -0.0062810761  0.0048474912  0.0075639477  8.838035e-05
#>  [9,] -0.0139645999 -0.0086715817 -0.0100768546 -7.384511e-03
#> [10,]  0.0086594872 -0.0044115611  0.0184038937  1.003820e-03
#> [11,]  0.0070770412  0.0118582470  0.0117563107  1.327666e-02
#> [12,] -0.0011375127  0.0012556871 -0.0065991893 -8.903156e-03
#> [13,] -0.0177420421 -0.0159321553 -0.0189994519 -1.088615e-02
#> [14,] -0.0049144447 -0.0009715927  0.0029291264 -2.867736e-03
#> [15,]  0.0026893119 -0.0070633630 -0.0023261220  4.115537e-03
#> [16,] -0.0115492409 -0.0126335930 -0.0052104045 -6.139118e-03
#> [17,]  0.0014166373  0.0024184787  0.0003302368  3.048938e-03
#> [18,]  0.0054385947 -0.0056420730 -0.0081907718 -8.511627e-04
#> [19,]  0.0171858988  0.0094823436  0.0094426624  6.704464e-03
#> [20,] -0.0093964423  0.0036357327 -0.0146397150 -1.830931e-03
#> 
#> $p
#>             [,1]
#>  [1,] 0.04531510
#>  [2,] 0.05199406
#>  [3,] 0.05238225
#>  [4,] 0.05235466
#>  [5,] 0.04947864
#>  [6,] 0.05492518
#>  [7,] 0.04862435
#>  [8,] 0.04937134
#>  [9,] 0.04735651
#> [10,] 0.04804812
#> [11,] 0.05536423
#> [12,] 0.04825235
#> [13,] 0.04462584
#> [14,] 0.04792000
#> [15,] 0.05070543
#> [16,] 0.04567733
#> [17,] 0.05159628
#> [18,] 0.05081562
#> [19,] 0.05297762
#> [20,] 0.05221507
#>