Panic-Copula for Stress-Testing

This function generates a large dimensional panic copula that can be further used for stress-testing and panic-aware portfolio optimization.

panic_copula(
  x,
  n = 10000,
  panic_cor = 0.99,
  panic_prob = 0.02,
  dist = c("normal", "t")
)

Arguments

x
A rectangular (non-tidy) data structure.

n
An integer with the number of scenarios to be generated.

panic_cor
A numeric value for the correlation in panic markets.

panic_prob
A numeric value with the probability in which panic markets can be triggered.

dist
A character with normal or t. The default is normal

Value

An object of the panic class.

Examples

x <- diff(log(EuStockMarkets))
panic_copula(x = x, n = 20, panic_cor = 0.99, panic_prob = 0.02)

#> $simulation
#>                 ..1           ..2           ..3           ..4
#>  [1,] -0.0090250282 -0.0081043744 -0.0078589327  0.0056694731
#>  [2,]  0.0058297733  0.0052350712  0.0062429396  0.0104180379
#>  [3,]  0.0018542793  0.0121098921 -0.0019856951  0.0069721250
#>  [4,] -0.0134855711 -0.0098751137 -0.0184038935 -0.0132766599
#>  [5,]  0.0044368808  0.0028026002  0.0117762944 -0.0024110531
#>  [6,]  0.0073388179 -0.0039842692  0.0096646481  0.0045036875
#>  [7,]  0.0109969231 -0.0016651219  0.0006587328  0.0004752131
#>  [8,]  0.0006151370 -0.0005523859 -0.0033421632 -0.0014324903
#>  [9,]  0.0031209746  0.0065901764  0.0144413175  0.0034276332
#> [10,]  0.0171858988  0.0154327448  0.0047513304  0.0084954771
#> [11,]  0.0090250284  0.0081043746  0.0078589329 -0.0056694729
#> [12,] -0.0058297730 -0.0052350710 -0.0062429393 -0.0104180377
#> [13,] -0.0018542791 -0.0121098920  0.0019856954 -0.0069721248
#> [14,]  0.0134855713  0.0098751139  0.0184038937  0.0132766600
#> [15,] -0.0044368806 -0.0028026000 -0.0117762942  0.0024110533
#> [16,] -0.0073388177  0.0039842694 -0.0096646479 -0.0045036873
#> [17,] -0.0109969229  0.0016651221 -0.0006587326 -0.0004752129
#> [18,] -0.0006151367  0.0005523861  0.0033421635  0.0014324904
#> [19,] -0.0031209743 -0.0065901762 -0.0144413172 -0.0034276331
#> [20,] -0.0171858987 -0.0154327448 -0.0047513301 -0.0084954769
#> 
#> $p
#>       [,1]
#>  [1,] 0.05
#>  [2,] 0.05
#>  [3,] 0.05
#>  [4,] 0.05
#>  [5,] 0.05
#>  [6,] 0.05
#>  [7,] 0.05
#>  [8,] 0.05
#>  [9,] 0.05
#> [10,] 0.05
#> [11,] 0.05
#> [12,] 0.05
#> [13,] 0.05
#> [14,] 0.05
#> [15,] 0.05
#> [16,] 0.05
#> [17,] 0.05
#> [18,] 0.05
#> [19,] 0.05
#> [20,] 0.05
#>